Restructure radiation callbacks.
